Try harder to find a valid dbus machine-id
Refactor the logic in ibus_get_local_machine_id() to try harder to find
a valid machine-id. Since the stable release of dbus 1.6.0,
/etc/machine-id is also a valid location for this file. If the user's
system is still not configured properly, fallback on a valid boot-id,
taken from the kernel. This avoids assigning an invalid (non-hex)
machine-id which may cause ibus frontends to not work properly.

If in the unlikely event that all 3 of these methods fail, then use the
invalid "machine-id" string to make a best effort of recovery.