Add IBusProperty.symbol for the short label.
IBusProperty.label had been shared between the panel menu label and
language-bar button label so the label is too short against the panel menu.
Now we redefines IBusProperty.label as the long label and the new
IBusProperty.symbol is defined as the short label for the language-bar
and the panel status icon.

TEST=Manually

Review URL: https://codereview.appspot.com/6495117