Notify engines of the content-type of input context
Add a new D-Bus method SetContentType to InputContext and Engine, to
notify engines of the content-type (primary purpose and hints) of
input context.  This is useful to implement intelligent features in
engines, such as automatic input-mode switch and text prediction.

The "purpose" and "hints" arguments are compatible with
GtkInputPurpose and GtkInputHints:
https://developer.gnome.org/gtk3/unstable/GtkEntry.html#GtkInputPurpose
https://developer.gnome.org/gtk3/unstable/GtkEntry.html#GtkInputHints

and the API is similar to the content_type event in the Wayland Input
Method Framework:
http://cgit.freedesktop.org/wayland/weston/tree/protocol/input-method.xml#n202

BUG=

Review URL: https://codereview.appspot.com/11422043