dconf: Work around using dbus development builds and /etc/machine-id
Recent DBus changed the way it reads /etc/machine-id to be more
strict, and it turns out that this breaks the use of dbus-launch here.

The *correct* fix is to use `dbus-run-session`, but not everyone has
that yet.  This is a quick hack that keeps the build going.

BUG=https://github.com/ibus/ibus/pull/16
TEST=data/dconf/00-upstream-settings

Review URL: https://codereview.appspot.com/209810043
Patch from Colin Walters <walters@verbum.org>.