Support vertical writing with IBUS_INPUT_HINT_VERTICAL_WRITING
GTK_INPUT_HINT_VERTICAL_WRITING has been supported in gtk 3.18
Put the vertical lookup table at the left top of the cursor position
in case of the vertical mode.
Also the gtk client tries to get the character width not to overlap
the character when the client window is moved to the screen left.

BUG=https://code.google.com/p/ibus/issues/detail?id=1780

Review URL: https://codereview.appspot.com/242140043