Enable to show a property on ibus panel icon
If an engine sets the value of icon-prop-key of IBusEngineDesc to
a property name of IBusProperty.key, the panel will show the engine
property in the icon with a delayed time.
This way can show both the engine icon and the engine mode in one icon.

Review URL: https://codereview.appspot.com/248950043