ibus-daemon update user cache when user observed path is updated
After the system registry cache is generated by root,
the cache remembers the observed path exists in /root == '~'
but if a user updates the user observed path, the observed path
exists in '~' and ibus-daemon cannot get the updated mtime
because the cache refers the root directory but not the user directory.
The patch tries to save '~' instead of the actual home dir
in the cache.

BUG=https://github.com/ibus/ibus/issues/1815
R=Shawn.P.Huang@gmail.com

Review URL: https://codereview.appspot.com/261040043