Add SetCursorLocationRelative D-Bus method
We have SetCursorLocation D-Bus method which takes the absolute
coordinates of the screen, while on Wayland the coordinates are relative
to the current input window and there is no way to calculate the
absolute coordinates.  That makes it impossible for the panel to
position the lookup table correctly.

This patch adds a new D-Bus method, SetCursorLocationRelative, which
takes the relative coordinates, so that the panel
implementation (typically the same process of Wayland compositor) can
use that information.  The counterpart of this changes is at:
https://bugzilla.gnome.org/show_bug.cgi?id=753476

Unlike SetCursorLocation, SetCursorLocationRelative is not delivered to
engines, since it is not very useful for the engines without knowing the
current input window (which is also not accessible from a different
process on Wayland).

Co-authored-by: Rui Matos <tiagomatos@gmail.com>

BUG=
R=takao.fujiwara1@gmail.com

Review URL: https://codereview.appspot.com/290780043 .