tests/ibus-engine-switch: Don't try to remove non-existing GSource
ibus-engine-switch schedules an idle doing an engine switch, and
also schedules a timeout after 1 microsecond which exits the main
loop. Both of these callbacks will return FALSE, which detaches
the associated GSource from the mainloop and destroys it.

As the only way to exit the mainloop is when the timeout callback runs,
we should not try to run g_source_remove(timeout_id) as it will already
be gone, and this causes a warning.

This commit reorders a bit to make the ordering between the idle and
timeout more obvious, and it gets rid of the calls to g_source_remove()
as they are needed.

BUG=https://github.com/ibus/ibus/issues/1868

Review URL: https://codereview.appspot.com/103100044