Fix GVariant leaks
The expectation is that g_dbus_message_set_body() takes ownership of the
GVariant, but this does not happen if the BusInputContext connection is NULL.
Call g_variant_unref() in that case to free the memory. Alternatively, a
GVariantBuilder could be used.

BUG=https://github.com/ibus/ibus/pull/1872
R=Shawn.P.Huang@gmail.com

Review URL: https://codereview.appspot.com/307050043

Patch from Eric R. Schulz <eric@ers35.com>.