ui/gtk3: Change modal dialog to focused dialog
There are several problems with the current emoji modal dialog.
If keyboard is grabbed on the popup window, the focus out/in events
cannot be detected so the dialog cannot be closed by the focus changes.

If mouse operation is supported on the popup window, need a custom
GtkHeaderBar with extended handle.vala but the behavior is unclear. [1]

Also current popup window has several extensions for the keyboard grab.

If the closed button is needed on the popup window since the focus events
cannot be detected, I think there is no merit to use the popup window.
Now IBusEmojier simply uses the focused window.

[1] https://mail.gnome.org/archives/gtk-app-devel-list/2017-April/msg00017.html

R=Shawn.P.Huang@gmail.com, alexepico@gmail.com

Review URL: https://codereview.appspot.com/316510043