Implement ibus-extension-gtk3 for the global keybinding
Currently IBus panel (ibus-ui-gtk3) is not available in GNOME and Plasma
so the emoji and unicode point typings are not available in GNOME and Plasma.
The workaround `ibus emoji` command is available but it put the selected
character into the copy buffer and users have to paste the character.

Originaly the emoji feature was implemented in IBus GtkIMModule but
it had several problems; the first is the keybinding is hard-coded
and IBus GtkIMModule does not use GSettings for the customized settings.
The second is the feature was available for GTK applications.
The third is that XKB input sources uses gtk-im-context-simple
but not ibus in GNOME desktop so users have to add an IM input sources
to enable IBus for the XKB input sources. The fourth is the feature
was available for IBusEngineSimple only and other IBus IMEs need to
inherit that class to get the emoji feature. The fifth is that
emoji typing is available for English only since IBusEngineSimple
had the feature. The sixth is that the default one dimension lookup
window was not useful to choose an emoji and needed two dimensions
lookup window.

And the implementation was moved from IBus GtkIMModule to IBus panel
to fix above problems.
But users have to use `ibus emoji` at present if ibus-ui-gtk3
is not available.

Now I think to move the emoji feature from ibus-ui-gtk3 to another
IBus component; ibus-extension-gtk3 which manages the Ctrl-Shift-e.
GNOME and Plasma desktops still do not show the GUI menu but
the shortcut key is available in this implementation.

BUG=RHBZ#1430501
R=Shawn.P.Huang@gmail.com

Review URL: https://codereview.appspot.com/339300043