ui/gtk3: Load Unicode data when open the dialog by default
The emoji data requires about 10MB and the Unicode data requires about 15MB.
Now the emoji data is loaded at the time of startup and the Unicode data
is loaded if users open the dialog at the beginning.
The settings can be customized with gsettings command and the keys
of 'load-emoji-at-startup' and 'load-unicode-at-startup' in
'org.freedesktop.ibus.panel.emoji' schema.

Review URL: https://codereview.appspot.com/340780043