panelbinding: Fix SEGV in panel_binding_parse_accelerator
panel_binding_parse_accelerator() could return NULL of the unowned
IBus.ProcessKeyEventData with gcc optimization.
Since Vala does not provice a static local variable, the variable is
moved to the class member to fix this SEGV.
Also fixed an infinite loop to show a compose preedit in xterm.
Also a NULL preedit is fixed in the first emoji candidate from the
emoji category window.