bus: Allow using non-abstract UNIX domain sockets
Previously, it was not possible to access ibus daemon from a separate
network namespace because its socket resided in an abstract namespace,
which is not shareable between network namespaces.

Passing --address=unix:path=/run/user/1000/ibus to ibus-daemon allows
clients running in containerized environment to communicate with daemon
as long as /run/user/1000 is mounted within container.

BUG=https://github.com/ibus/ibus/pull/2077