ui/gtk3: Fix to exit `ibus emoji` with Escape key
When Escape key is pressed, the current ibus-ui-emojier should exit
although another daemon one is alive.

`ibus emoji` was used in Wayland desktop as a workaround but now
IBus emoji functions are moved to the engine's pre-edit and the same
function is available with Ctrl-Sfhit-e and space key so this problem
would not be so visible.

BUG=rhbz#1658187