src/tests: Run gnome-session with no-overview mode
gnome-shell 40 now shows the overview mode by login and
gnome-desktop-testing-runner cannot get the input focus.
Running gnome-session with no-overview shell extension can
get the input focus or mutter is another option.
The default may be changed to twm if mutter also will be changed
not to accept the application focus in the future.

Disable Tour dialog which prevent test application from getting the
input focus.

Don't output FAIL if the actual failure is 0 for Fedora CI.

BUG=https://discourse.gnome.org/t/focus-on-autostart-application-by-login/5863