setup: Enhance engine search function
ibus-setup can search both the language names and input method names
in the top language list but when you search a language keyword
in the language list and move to the input method list after click
the hit language name, any input methods could not be shown because
the language didn't hit in any input method names.

In this enhancement, ibus-setup can show the input methods to hit
the language names.