Fix unref problems with floating references
When running with debug-enabled GLIB there are several critical
errors output: "A floating object ... was finalized. This means
that someone called g_object_unref() on an object that had only
a floating reference; the initial floating reference is not
owned by anyone and must be removed with g_object_ref_sink()."

This change fixes this by calling `g_object_ref_sink()` before
`g_object_unref()` if we have a floating reference.

It also fixes another related problem where we called
`g_object_unref()` on a static IBusText string (created with
`ibus_text_new_from_static_string()`) for which the API documentation
says not to free.

BUG=https://github.com/ibus/ibus/pull/2359