src/ibuscompose: Check algorithm dead key by 9 length
This change fixes the D-Bus timeout by a long compose preedit text.
check_normalize_nfc() calculates the factorical of the compse preedit
length and assigne the value to an INT variable and it could excceed
the MAX_INT.
Probably I think the length of the compose algorizhm dead key would be
enough 9. I.e. 9! == 46320 <= SHRT_MAX == 32767

BUG=https://github.com/ibus/ibus/issues/2385