ui/gtk3: Disable XKB engines in Plasma Wayland
Currently ibus-ui-gtk3 runs the setxkbmap command internally to set
the XKB keymaps from XKB engines but setxkbmap does not work in Wayland
session.
The IBus XKB engines are disabled at the moment in Plamsa Wayland
and ibus-ui-gtk3 asks users to use systemsettings5.

Will use libinput and libxkbcommon later but it would be better
to implement IBus in Plamsa Wayland compositor.

BUG=https://github.com/ibus/ibus/issues/2408