src: Add IBUS_CAP_OSK to IBusCapabilite
Some IMEs' behavior is different between the on-screen keyboard and
the direct physical keyboard and this flag is useful for the IMEs.

Also fix src/ibusaccelgroup.c for gtkdoc-mkhtml.
If the API comment of IBusCapabilite is updated, XML & HTML files
are rebuilt and gtk-doc-1.33.2 no longer accepts HTML tags in
the comments.